add fast mode

+++ PureseqTM.sh
@@ -4,11 +4,11 @@
 # ----- usage ------ #
 usage()
 {
-	echo "PureseqTM v0.19 [Apr-30-2019] "
+	echo "PureseqTM v0.20 [May-03-2019] "
 	echo "    Predict TransMembrane (TM) topology from a given sequence in FASTA format. "
 	echo ""
-	echo "USAGE:  ./PureseqTM.sh <-i input_fasta | input_tgt> [-o out_root] [-p signal_pep] [-P plot] [-K remove_tmp] [-H home] "
-	echo "                       [-l truth_label] [-s high_thres] [-S low_thres] [-c high_len] [-C low_len] [-L seg_len] "
+	echo "USAGE:  ./PureseqTM.sh <-i input_fasta | input_tgt> [-o out_root] [-m mode] [-p signal_pep] [-P plot] [-K remove_tmp] "
+	echo "                       [-l truth_label] [-s high_thres] [-S low_thres] [-c high_len] [-C low_len] [-L seg_len] [-H home] "
 	echo "Options:"
 	echo ""
 	echo "***** required arguments *****"
@@ -19,14 +19,15 @@ usage()
 	echo "***** optional arguments *****"
 	echo "-o out_root     : Default output would the current directory. [default = './\${input_name}_PureTM'] "
 	echo ""
+	echo "-m mode         : Fast or slow mode. [default = 1 use slow mode] "
+	echo "                  set 0 to use fast mode that DO NOT run phobius and philius "
+	echo ""
 	echo "-p signal_pep   : Output signal peptide label or not. [default = 0 NOT to output] "
 	echo ""
 	echo "-P plot         : Plot posterior probabilities or not. [default = 0 NOT to plot] "
 	echo ""
 	echo "-K remove_tmp   : Remove temporary folder or not. [default = 1 to remove] "
 	echo ""
-	echo "-H home         : home directory of PureseqTM_Package. "
-	echo "                  [default = `dirname $0`] "
 	echo "***** additional arguments *****"
 	echo "-l truth_label  : The ground-truth 2-state TM label (with 1 for TM and 0 for non-TM) in FASTA format "
 	echo "                  [default = null] "
@@ -39,6 +40,10 @@ usage()
 	echo "-C low_len      : The low cutoff length to detect 'TM' regions. [default = 10] "
 	echo ""
 	echo "-L seg_len      : The segment len cutoff for 'cut' operation. [default = 30] "
+	echo ""
+	echo "***** home directory *****"
+	echo "-H home         : home directory of PureseqTM_Package. "
+	echo "                  [default = `dirname $0`] "
 	echo ""
 	exit 1
 }
@@ -71,7 +76,8 @@ input_tgt=""
 amino_only=0
 #-> others
 out_root=""
-plot=0              #-> default :NOT plot posterior probabilities
+mode=1              #-> default: RUN with slow mode (run Phobius and Philius)
+plot=0              #-> default: NOT plot posterior probabilities
 signal_pep=0        #-> default: NOT output signal peptide
 kill_tmp=1          #-> default: kill temporary root
 home=`dirname $0`   #-> default: dirname of the 0-th argument
@@ -85,7 +91,7 @@ seg_len="30"
 
 
 #-> parse arguments
-while getopts ":i:o:p:P:K:H:l:s:S:c:C:L:" opt;
+while getopts ":i:o:m:p:P:K:H:l:s:S:c:C:L:" opt;
 do
 	case $opt in
 	#-> required arguments
@@ -95,6 +101,9 @@ do
 	#-> optional arguments
 	o)
 		out_root=$OPTARG
+		;;
+	m)
+		mode=$OPTARG
 		;;
 	p)
 		signal_pep=$OPTARG
@@ -237,10 +246,6 @@ fi
 #----------- Part 1.1 generate Phobius features ---------------#
 params=$param/phobius
 runcom="$bin/decodeanhmm -f $params/phobius.options $params/phobius.model"
-$runcom $tmp_root/$relnam.seq -plp 1> $tmp_root/$relnam.plp 2> $tmp_root/$relnam.err
-grep -v "#" $tmp_root/$relnam.plp | \
-            awk '{a=$2;b=$3+$4;c=$5;d=$6+$7+$8+$9;print a" "b" "c" "d}' \
-            > $tmp_root/$relnam.phobius
 #-> get the number of TM segment from Phobius
 $runcom -GFF $tmp_root/$relnam.seq 2> $tmp_root/$relnam.err | grep -v "SEQUENCE" \
              | grep -v "#" > $tmp_root/$relnam.pbseg
@@ -249,23 +254,38 @@ tmseg=`awk '{if($3=="M"){print $0}}' $tmp_root/$relnam.pbseg | wc | awk '{print 
 #-> check the existance of N-termi signal peptide from Phobius
 ntsp=`awk '{if($3=="h"){print $0}}' $tmp_root/$relnam.pbseg | wc | awk '{print $1}'`
 
+#-> run Phobius to generate HMM features
+if [ $mode -eq 1 ]
+then
+	$runcom $tmp_root/$relnam.seq -plp 1> $tmp_root/$relnam.plp 2> $tmp_root/$relnam.err
+	grep -v "#" $tmp_root/$relnam.plp | \
+	    awk '{a=$2;b=$3+$4;c=$5;d=$6+$7+$8+$9;print a" "b" "c" "d}' \
+	    > $tmp_root/$relnam.phobius
+fi
+
+
 #----------- Part 1.2 generate Philius features ---------------#
-params=$param/philius
-$util/AA_to_CODE $tmp_root/$relnam.seq $tmp_root/$relnam.obs
-echo "$tmp_root/$relnam.obs" > $tmp_root/$relnam.list
-$bin/gmtkJT -strFile $params/model_test1.str \
-            -triFile $params/model_test1.mytrifile \
-            -of1 $tmp_root/$relnam.list -nf1 0 -ni1 2 -fmt1 ascii \
-            -inputMasterFile $params/model_test1.master.Viterbi \
-            -inputTrainableParameters $params/learnedParamsALL.out \
-            -pCliquePrintRange 1:2 \
-            -cCliquePrintRange 1:1 \
-            -eCliquePrintRange 1:2 \
-            -doDist \
-            -cptNormThreshold 10. \
-            -verbosity 1 > $tmp_root/$relnam.runJT
-rm -f jt_info.txt
-$util/Proc_runJT $tmp_root/$relnam.runJT $tmp_root/$relnam.philius
+
+#-> run Philius to generate DBN features
+if [ $mode -eq 1 ]
+then
+	params=$param/philius
+	$util/AA_to_CODE $tmp_root/$relnam.seq $tmp_root/$relnam.obs
+	echo "$tmp_root/$relnam.obs" > $tmp_root/$relnam.list
+	$bin/gmtkJT -strFile $params/model_test1.str \
+	    -triFile $params/model_test1.mytrifile \
+	    -of1 $tmp_root/$relnam.list -nf1 0 -ni1 2 -fmt1 ascii \
+	    -inputMasterFile $params/model_test1.master.Viterbi \
+	    -inputTrainableParameters $params/learnedParamsALL.out \
+	    -pCliquePrintRange 1:2 \
+	    -cCliquePrintRange 1:1 \
+	    -eCliquePrintRange 1:2 \
+	    -doDist \
+	    -cptNormThreshold 10. \
+	    -verbosity 1 > $tmp_root/$relnam.runJT
+	rm -f jt_info.txt
+	$util/Proc_runJT $tmp_root/$relnam.runJT $tmp_root/$relnam.philius
+fi
 
 
 #--------------------------------------------------------------#
@@ -273,8 +293,13 @@ $util/Proc_runJT $tmp_root/$relnam.runJT $tmp_root/$relnam.philius
 #--------------------------------------------------------------#
 
 $util/pureseq_dump $tmp_root/$relnam.seq 1 > $tmp_root/$relnam.pureseq
-paste $tmp_root/$relnam.phobius $tmp_root/$relnam.philius \
-            $tmp_root/$relnam.pureseq > $tmp_root/$relnam.feat_prev
+if [ $mode -eq 1 ]
+then
+	paste $tmp_root/$relnam.phobius $tmp_root/$relnam.philius \
+	    $tmp_root/$relnam.pureseq > $tmp_root/$relnam.feat_prev
+else
+	cp $tmp_root/$relnam.pureseq $tmp_root/$relnam.feat_prev
+fi
 $bin/DeepCNF_FeatMake $tmp_root/$relnam.feat_prev $truth_label_file \
             > $tmp_root/$relnam.feat
 
@@ -304,20 +329,37 @@ fi
 #--------------------------------------------------------------#
 
 labeldim="2"
-featdim="36"
+featdim_fast="25"
+featdim_slow="36"
 profdim="65"
 wind="5,5,5,5,5"
 node="100,100,100,100,100"
-detect_model=$param/detect.model
-predict_model=$param/puretm.model
+detect_model_fast=$param/detect.model_fast
+predict_model_fast=$param/puretm.model_fast
+detect_model_slow=$param/detect.model_slow
+predict_model_slow=$param/puretm.model_slow
 profile_model=$param/proftm.model
 addi=""
-#---- discriminate TM and nonTM -----#
+
+#---- use fast or slow model --------#
+if [ $mode -eq 1 ]
+then
+	featdim=$featdim_slow
+	detect_model=$detect_model_slow
+	predict_model=$predict_model_slow
+else
+	featdim=$featdim_fast
+	detect_model=$detect_model_fast
+	predict_model=$predict_model_fast
+fi
+
+#---- discriminate TP and nonTP -----#
 $bin/DeepCNF_Pred -i $tmp_root/$relnam.feat -w $wind -d $node -s $labeldim -l $featdim \
             -m $detect_model $addi > $tmp_root/$relnam.detect_out 2> $tmp_root/$relnam.err2
 grep -v "#" $tmp_root/$relnam.detect_out | awk '{if($4>thres){print $0}}' thres=$high_thres \
             > $tmp_root/$relnam.detect_reso
 rm -f $tmp_root/$relnam.err2
+
 #---- if TM region is detected, then run predict -----#
 len=`wc $tmp_root/$relnam.detect_reso | awk '{print $1}'`
 if [ $len -gt $high_len ]
@@ -334,6 +376,7 @@ else
 	cp $tmp_root/$relnam.detect_out $tmp_root/$relnam.out
 fi
 rm -f $tmp_root/$relnam.err3
+
 
 #---- make output readable -----#
 $util/Verify_FASTA $tmp_root/$relnam.seq $out_root/$relnam.fasta
@@ -388,6 +431,6 @@ else
 fi
 
 #=============== exit ===============#
-exit
+exit 0
